+++ packages/solid_generator/test/golden/outputs/effect_with_signal_and_computed.g.dart
@@ -14,14 +14,14 @@ class _EffectWithDepsState extends State<EffectWithDeps> {
     () => counter.value * 2,
     name: 'doubleCounter',
   );
-  late final logBoth = Effect(() {
-    print('${counter.value} / ${doubleCounter.value}');
-  }, name: 'logBoth');
+  late final Effect logBoth;
 
   @override
   void initState() {
     super.initState();
-    logBoth;
+    logBoth = Effect(() {
+      print('${counter.value} / ${doubleCounter.value}');
+    }, name: 'logBoth');
   }
 
   @override